+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.4.1
 - fix icon visualisation in pop-up and improve pop-up format
+- support label for legends
 
 ## v1.4.0
 - Migrate to react for compatibility with Grafana 7+